I'm not sure what the best way to make a cache file is, taking a break on this for now. Anyone else feel free to modify my code.

+++ main/romcache.h
@@ -26,7 +26,7 @@ typedef struct {
     char MD5[33];
     char goodname[256];
     char path[PATH_MAX];
-    unsigned long os_timestamp;
+    int os_timestamp;
 } cache_entry;
 
 typedef struct {
